Use (float) typecast to insure ratios between 50-99% are calculated correctly (not rounded up to 100%).

+++ src/sbbs3/chk_ar.cpp
@@ -394,7 +394,7 @@ bool sbbs_t::ar_exp(uchar **ptrptr, user_t* user)
 				break;
 			case AR_PCR:	/* post/call ratio (by percentage) */
 				if(user->logons>user->posts
-					&& (!user->posts || (100/(user->logons/user->posts))<(long)n))
+					&& (!user->posts || (100/((float)user->logons/user->posts))<(long)n))
 					result=_not;
 				else
 					result=!_not;
@@ -406,7 +406,7 @@ bool sbbs_t::ar_exp(uchar **ptrptr, user_t* user)
 				l=user->dlb;
 				if(!l) l=1;
 				if(user->dlb>user->ulb
-					&& (!user->ulb || (100/(l/user->ulb))<n))
+					&& (!user->ulb || (100/((float)l/user->ulb))<n))
 					result=_not;
 				else
 					result=!_not;
@@ -418,7 +418,7 @@ bool sbbs_t::ar_exp(uchar **ptrptr, user_t* user)
 				i=user->dls;
 				if(!i) i=1;
 				if(user->dls>user->uls
-					&& (!user->uls || (100/(i/user->uls))<n))
+					&& (!user->uls || (100/((float)i/user->uls))<n))
 					result=_not;
 				else
 					result=!_not;
